implement ls

+++ kernel/include/stddef.h
@@ -13,7 +13,7 @@ typedef unsigned int off_t;
 typedef unsigned int time_t;
 typedef int pid_t;
 typedef unsigned short dev_t;
-typedef uint32_t ino_t;
+typedef uint16_t ino_t;
 typedef uint32_t uid_t;
 typedef uint32_t gid_t;
