+++ libc/include/sys/types.h
@@ -15,7 +15,7 @@ typedef int ssize_t;
 typedef unsigned int off_t;
 typedef unsigned int time_t;
 typedef int pid_t;
-typedef uint32_t ino_t;
+typedef uint16_t ino_t;
 typedef uint32_t uid_t;
 typedef uint32_t gid_t;
